requirements.txt
opencv-python
numpy
pillow
pytest
enum34
configparser
